+++ android/src/main/kotlin/com/mapbox/maps/mapbox_maps/mapping/LogoMappings.kt
@@ -2,23 +2,25 @@
 package com.mapbox.maps.mapbox_maps.mapping
 
 import android.content.Context
+import com.mapbox.maps.mapbox_maps.toDevicePixels
+import com.mapbox.maps.mapbox_maps.toLogicalPixels
 import com.mapbox.maps.pigeons.FLTSettings
 import com.mapbox.maps.plugin.logo.generated.LogoSettingsInterface
 
 fun LogoSettingsInterface.applyFromFLT(settings: FLTSettings.LogoSettings, context: Context) {
   settings.position?.let { position = it.toPosition() }
-  settings.marginLeft?.let { marginLeft = it.toFloat() }
-  settings.marginTop?.let { marginTop = it.toFloat() }
-  settings.marginRight?.let { marginRight = it.toFloat() }
-  settings.marginBottom?.let { marginBottom = it.toFloat() }
+  settings.marginLeft?.let { marginLeft = it.toDevicePixels(context) }
+  settings.marginTop?.let { marginTop = it.toDevicePixels(context) }
+  settings.marginRight?.let { marginRight = it.toDevicePixels(context) }
+  settings.marginBottom?.let { marginBottom = it.toDevicePixels(context) }
 }
 
-fun LogoSettingsInterface.toFLT() = FLTSettings.LogoSettings.Builder().let { settings ->
+fun LogoSettingsInterface.toFLT(context: Context) = FLTSettings.LogoSettings.Builder().let { settings ->
   settings.setPosition(position.toOrnamentPosition())
-  settings.setMarginLeft(marginLeft.toDouble())
-  settings.setMarginTop(marginTop.toDouble())
-  settings.setMarginRight(marginRight.toDouble())
-  settings.setMarginBottom(marginBottom.toDouble())
+  settings.setMarginLeft(marginLeft.toLogicalPixels(context))
+  settings.setMarginTop(marginTop.toLogicalPixels(context))
+  settings.setMarginRight(marginRight.toLogicalPixels(context))
+  settings.setMarginBottom(marginBottom.toLogicalPixels(context))
   settings.build()
 }