+++ example/integration_test/camera_test.dart
@@ -429,19 +429,7 @@ void main() {
     await tester.pumpAndSettle();
     final mapboxMap = await mapFuture;
 
-    var destination = ScreenCoordinate(x: 100, y: 100);
-
-    // the input has to be scaled on Android,
-    // as GL Native treats iOS and Android coordinates differently:
-    // * on iOS screen coordinates are expected to be provided(as well as returned by GL Native) as logical pixels
-    // * on Android screen coordinates are expected to be physical pixels(both ways)
-    // TODO: this should be removed once https://mapbox.atlassian.net/browse/MAPSFLT-120 is addressed
-    if (Platform.isAndroid) {
-      final BuildContext context = tester.element(find.byType(MapWidget));
-      final pixelRatio = MediaQuery.of(context).devicePixelRatio;
-      destination.x = destination.x * pixelRatio;
-      destination.y = destination.y * pixelRatio;
-    }
+    final destination = ScreenCoordinate(x: 100, y: 100);
     final options = await mapboxMap.getDragCameraOptions(
         ScreenCoordinate(x: 0, y: 0), destination);
     final coordinates = options.center!["coordinates"] as List;